+++ src/decoder.h
@@ -14,7 +14,7 @@ public:
     Decoder &operator=(const Decoder &) = delete;
 
     // decode varint
-    static uint decode_varint(boost::container::vector<char> &data);
+    static uint64_t decode_varint(boost::container::vector<char> &data);
 
     static void decode_message(Buffer_t &buf, Profile_t &prof, boost::container::vector<char> &data);
